absent

+++ lib/view/student/dashboard/section/metadata/view.dart
@@ -1,6 +1,6 @@
 import 'package:attendance_nmsct/view/student/dashboard/section/metadata/metadata.dart';
-import 'package:attendance_nmsct/widgets/else_statement.dart';
 import 'package:firebase_storage/firebase_storage.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:loader_skeleton/loader_skeleton.dart';
@@ -16,6 +16,8 @@ class MetaDataIndex extends StatefulWidget {
 }
 
 class _MetaDataIndexState extends State<MetaDataIndex> {
+  final GlobalKey<RefreshIndicatorState> _refreshIndicatorKey =
+      GlobalKey<RefreshIndicatorState>();
   List<Reference> _imageReferences = [];
   List _imageUrls = [];
 
@@ -38,6 +40,7 @@ class _MetaDataIndexState extends State<MetaDataIndex> {
       //     date)); // You may need to adjust the condition based on your file naming convention
       setState(() {
         _imageReferences = items.toList();
+
         _imageUrls = _imageReferences.map((ref) {
           return ref.getDownloadURL();
         }).toList();
@@ -51,11 +54,40 @@ class _MetaDataIndexState extends State<MetaDataIndex> {
   }
 
   Future<void> deleteImage(Reference imageRef) async {
-    try {
-      await imageRef.delete();
-      _getImageReferences(); // Refresh the list after deletion
-    } catch (e) {
-      print('Error deleting file: $e');
+    // Show a confirmation dialog
+    bool deleteConfirmed = await showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return CupertinoAlertDialog(
+          title: const Text('Confirm Deletion'),
+          content: const Text('Are you sure you want to delete this image?'),
+          actions: [
+            TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(false); // User canceled deletion
+              },
+              child: const Text('Cancel'),
+            ),
+            TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(true); // User confirmed deletion
+              },
+              child: const Text('Delete'),
+            ),
+          ],
+        );
+      },
+    );
+
+    if (deleteConfirmed == true) {
+      // User confirmed deletion
+      try {
+        await imageRef.delete();
+
+        _getImageReferences(); // Refresh the list after deletion
+      } catch (e) {
+        print('Error deleting file: $e');
+      }
     }
   }
 
@@ -71,87 +103,95 @@ class _MetaDataIndexState extends State<MetaDataIndex> {
   Widget build(BuildContext context) {
     screenHeight = MediaQuery.of(context).size.height;
     screenWidth = MediaQuery.of(context).size.width;
-    return Column(
-      children: [
-        if (isLoading)
-          Expanded(
-            child: CardPageSkeleton(),
-          )
-        else if (_imageReferences.isEmpty)
-          const Expanded(
-            child: Center(
-              child: Text(
-                'No data available.',
-                style: TextStyle(fontSize: 18),
+    return RefreshIndicator(
+      key: _refreshIndicatorKey,
+      onRefresh: _getImageReferences,
+      child: Column(
+        children: [
+          if (isLoading)
+            Expanded(
+              child: CardPageSkeleton(),
+            )
+          else if (_imageReferences.isEmpty)
+            const Expanded(
+              child: Center(
+                child: Text(
+                  'No data available',
+                  style: TextStyle(fontSize: 18),
+                ),
               ),
-            ),
-          )
-        else
-          Expanded(
-            child: Padding(
-              padding: const EdgeInsets.all(10),
-              child: ListView.builder(
-                itemCount: _imageReferences.length,
-                itemBuilder: (context, index) {
-                  final imageRef = _imageReferences[index];
-                  final imageName = imageRef.name;
+            )
+          else
+            Expanded(
+              child: Padding(
+                padding: const EdgeInsets.all(10),
+                child: ListView.builder(
+                  itemCount: _imageReferences.length,
+                  itemBuilder: (context, index) {
+                    final imageRef = _imageReferences[index];
 
-                  return Container(
-                    padding: EdgeInsets.only(
-                        bottom:
-                            index == _imageReferences.length - 1 ? 70.0 : 0),
-                    child: TimelineTile(
-                      isFirst: index == 0,
-                      isLast: index == _imageReferences.length - 1,
-                      alignment: TimelineAlign.start,
-                      indicatorStyle: const IndicatorStyle(
-                        width: 20,
-                        color: Colors.blue, // Adjust color as needed
-                      ),
-                      endChild: Container(
-                        constraints: BoxConstraints(
-                          maxWidth: MediaQuery.of(context).size.width * 0.7,
+                    final imageName = imageRef.name;
+
+                    return Container(
+                      padding: EdgeInsets.only(
+                          bottom:
+                              index == _imageReferences.length - 1 ? 70.0 : 0),
+                      child: TimelineTile(
+                        isFirst: index == 0,
+                        isLast: index == _imageReferences.length - 1,
+                        alignment: TimelineAlign.start,
+                        indicatorStyle: const IndicatorStyle(
+                          width: 20,
+                          color: Colors.blue, // Adjust color as needed
                         ),
-                        child: Card(
-                          child: ListTile(
-                            leading: FutureBuilder(
-                              future: _imageUrls[index],
-                              builder: (context, snapshot) {
-                                if (snapshot.connectionState ==
-                                    ConnectionState.done) {
-                                  return Image.network(
-                                    snapshot.data.toString(),
-                                    width: 50, // Adjust width as needed
-                                    height: 50, // Adjust height as needed
-                                  );
-                                } else {
-                                  return SizedBox.shrink();
-                                }
-                              },
-                            ),
-                            title: Text(
-                              imageName,
-                              style: TextStyle(fontSize: screenWidth / 25),
-                            ),
-                            onTap: () {
-                              Navigator.push(
-                                context,
-                                MaterialPageRoute(
-                                  builder: (context) =>
-                                      Meta_Data(image: imageRef),
+                        endChild: Container(
+                          constraints: BoxConstraints(
+                            maxWidth: MediaQuery.of(context).size.width * 0.7,
+                          ),
+                          child: GestureDetector(
+                            onLongPress: () => deleteImage(imageRef),
+                            child: Card(
+                              child: ListTile(
+                                leading: FutureBuilder(
+                                  future: _imageUrls[index],
+                                  builder: (context, snapshot) {
+                                    if (snapshot.connectionState ==
+                                        ConnectionState.done) {
+                                      return Image.network(
+                                        snapshot.data.toString(),
+                                        width: 50, // Adjust width as needed
+                                        height: 50, // Adjust height as needed
+                                      );
+                                    } else {
+                                      return const SizedBox.shrink();
+                                    }
+                                  },
                                 ),
-                              );
-                            },
+                                title: Text(
+                                  imageName,
+                                  style: TextStyle(fontSize: screenWidth / 25),
+                                ),
+                                onTap: () {
+                                  Navigator.push(
+                                    context,
+                                    MaterialPageRoute(
+                                      builder: (context) =>
+                                          Meta_Data(image: imageRef),
+                                    ),
+                                  );
+                                },
+                              ),
+                            ),
                           ),
                         ),
                       ),
-                    ),
-                  );
-                },
+                    );
+                  },
+                ),
               ),
             ),
-          ),
-      ],
+        ],
+      ),
     );
   }
 }
